CLS-23: Preliminary Refactoring of Testing Tools
